requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu130
torch==2.11.0+cu130
torchvision==0.26.0+cu130
torchaudio==2.11.0+cu130
accelerate>=1.0
openvino==2026.1.0
optimum[openvino]>=1.17.0
gradio>=4.0
transformers>=4.57.0,<5
pyannote.audio==4.0.4
pedalboard
noisereduce>=3.0
librosa
soundfile
python-dotenv
